requirements.txt
requests==2.31.0
pandas==1.5.3
numpy==1.23.5
google-cloud-storage==2.13.0
